chore(snownet): add more logging for connections (#9695)
In a recent release, `str0m` downgraded all INFO logs to DEBUG. Whilst
generally appreciated, it means we don't have a lot of visibility
anymore into which candidates are being exchanged and what the ICE
credentials of the connections are.

We re-add this information to our existing logs when creating and
updating connections.